requirements-test.txt
-r requirements.txt

coverage
coveralls
mock>=1.0.1
nose>=1.3.0
django-nose>=1.2
flake8>=2.1.0
tox>=1.7.0

pytest==3.6.3
pytest-django==3.3.2
pytest-cov==2.5.1
